♻️ refactor(monitoring): extract health status bar
Move the monitoring health status bar into a dedicated component and keep shared time-window formatting in the model layer.

This reduces MonitoringCenterPage coupling without changing rendering behavior; validated with type-check, targeted tests, and lint.
