+++ src/features/monitoring/model/statusWindow.ts
@@ -1,0 +1,19 @@
+export const formatStatusWindowLabel = (
+  startTime: number,
+  endTime: number,
+  locale: string
+) => {
+  const start = new Date(startTime);
+  const end = new Date(endTime);
+  const sameDay = start.toDateString() === end.toDateString();
+  const dateOptions: Intl.DateTimeFormatOptions = { month: 'numeric', day: 'numeric' };
+  const timeOptions: Intl.DateTimeFormatOptions = { hour: '2-digit', minute: '2-digit' };
+  const startDateLabel = start.toLocaleDateString(locale, dateOptions);
+  const endDateLabel = end.toLocaleDateString(locale, dateOptions);
+  const startTimeLabel = start.toLocaleTimeString(locale, timeOptions);
+  const endTimeLabel = end.toLocaleTimeString(locale, timeOptions);
+
+  return sameDay
+    ? `${startDateLabel} ${startTimeLabel} - ${endTimeLabel}`
+    : `${startDateLabel} ${startTimeLabel} - ${endDateLabel} ${endTimeLabel}`;
+};
